requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
openai-whisper==20231117
transformers==4.35.2
torch
pandas==2.1.3
python-multipart==0.0.6
nltk==3.8.1
